requirements.txt
asgiref==3.7.2
black==23.7.0
click==8.1.5
Django==4.2.3
django-extensions==3.2.3
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
pytz==2023.3
sqlparse==0.4.4
